Ports binary elementwise operations to TRT dialect
